% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Eugenia_aurata}
\alias{Eugenia_aurata}
\title{Eugenia aurata occurrence data.}
\format{An object of class \code{data.frame} with 286 rows and 4 columns.}
\usage{
Eugenia_aurata
}
\description{
Distribution data for Eugenia aurata. Downloaded from Global Biodiversity Information Facility(GBIF).
}
\references{
\url{http://www.gbif.org}
}
\keyword{datasets}
