% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Rasetr South America.}
\format{An object of class \code{RasterLayer} of dimension 303 x 317 x 1.}
\usage{
example
}
\description{
raster.
}
\references{
\url{http://www.gbif.org}
}
\keyword{datasets}
