% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{filt}
\alias{filt}
\title{Mark occurrences with the name informed different from the coordinate in a user supplied vector file}
\usage{
filt(pts, shape = NULL, field.pts = "municipality",
  field.shape = "NOMEMUNICP")
}
\arguments{
\item{pts}{data.frame. Table with points of occurrence, including the municipalities informed on the label.}

\item{shape}{It can be a shape of municipalities of Brazil in format "SpatialPolygonsDataFrame".}

\item{field.pts}{string. Name of pts column containing the standard to be searched for.}

\item{field.shape}{string.Name of shape column containing the standard to be searched for.}
}
\value{
a data frame
}
\description{
A function to check whether the coordinates are consistent with the name of the polygon.
}
\details{
empty yet
}
\examples{

filt(Eugenia_aurata, shape = world, field.pts = "country", field.shape = "LONG_NAME")

}
\author{
Diogo S. B. Rocha
}
