% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Leucaena_leucocephala}
\alias{Leucaena_leucocephala}
\title{Leucena leucocephala occurrence data.}
\format{An object of class \code{data.frame} with 6980 rows and 5 columns.}
\usage{
Leucaena_leucocephala
}
\description{
Distribution data for Leucena leucocephala. Downloaded from Global Biodiversity Information Facility(GBIF).
}
\references{
\url{http://www.gbif.org}
}
\keyword{datasets}
