% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.dups.R
\name{remove.dups}
\alias{remove.dups}
\title{Remove duplicated and NAs}
\usage{
remove.dups(pts, mask = example, na.rm = FALSE, dup = T)
}
\arguments{
\item{pts}{Data.frame. Table with points of occurrence, including the municipalities informed on the label.}

\item{mask}{String.Name of shape column containing the standard to be searched for.}

\item{na.rm}{String. Name of pts column containing the standard to be searched for.}

\item{dup}{logical}
}
\value{
a data frame
}
\description{
A function to check whether the coordinates are consistent with the name of the polygon.
}
\details{
The
}
\examples{

eug <- remove.dups(Eugenia_aurata, mask = example, na.rm = FALSE, dup = F)
eug

}
\author{
Diogo S. B. Rocha
}
