% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acentos.R
\name{rm_accent}
\alias{rm_accent}
\title{Remove words accents}
\usage{
rm_accent(str, pattern = "all")
}
\arguments{
\item{str}{vector of strings that will have their accents removed}

\item{pattern}{vector of strings with one or more elements indicating which accents should be removed.}
}
\value{
vector of strings with their accents removed
}
\description{
A function to remove words accents.
}
\details{
Originally available from Athos in \url{https://pt.stackoverflow.com/questions/46473/remover-acentos}
}
\examples{
inserir exemplo

}
\author{
Athos
}
