% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{World map.}
\format{An object of class \code{SpatialPolygonsDataFrame} with 251 rows and 2 columns.}
\usage{
world
}
\description{
shape.
}
\references{
\url{http://www.gbif.org}
}
\keyword{datasets}
