% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_mun.R
\docType{data}
\name{br_mun}
\alias{br_mun}
\title{Shapefile of the municipalities of Brazil.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
br_mun
}
\description{
Shapefile of the municipalities of Brazil.
}
\keyword{datasets}
