% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_mun.R
\docType{data}
\name{br_mun}
\alias{br_mun}
\title{Shapefile of the municipalities of Brazil.}
\format{A SpatialPolygonsDataFrame with 1111111 municipalities and 9999 variables:
\describe{
  \item{area}{price, in US dollars}
  \item{perimeter}{weight of the diamond, in carats}
  ...
}}
\source{
\url{https://downloads.ibge.gov.br/}
}
\usage{
br_mun
}
\description{
Shapefile of the municipalities of Brazil.
}
\keyword{datasets}
