% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br_mun}
\alias{br_mun}
\title{Shapefile of the municipalities of Brazil.}
\format{A SpatialPolygonsDataFrame with 5907 municipalities and 21 variables:
\describe{
  \item{area}{area of the municipality}
  \item{perimeter}{perimeter of the municipality}
  \item{NOMEMUNICP}{name of the municipality}
  ...
}}
\source{
\url{https://downloads.ibge.gov.br/}
}
\usage{
br_mun
}
\description{
An object of the type SpatialPolygonsDataFrame created from the shapefile of municipalities in Brazil, from the Brazilian Institute of Geography and Statistics or IBGE (Portuguese: Brazilian Institute of Geography and Statistics)
}
\keyword{datasets}
