% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euterpe.R
\docType{data}
\name{euterpe}
\alias{euterpe}
\title{Euterpe edulis data.}
\format{Distribution data for Euterpe edulis (a widely distributed palm, occurring continuously throughout the Atlantic Forest). Downloaded from GBIF with the gbif function from dismo package.}
\usage{
euterpe
}
\description{
Euterpe edulis data.
}
\references{
\url{http://www.gbif.org}
}
\keyword{datasets}
