% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtro_AM.R
\name{filt.generic}
\alias{filt.generic}
\title{Marks occurrences based on user supplied vector file}
\usage{
filt.generic(pts, shape = NULL, few.pts = F, value = 10,
  plot = TRUE, file = NULL)
}
\arguments{
\item{pts}{Data.frame. Table with points of occurrence.}

\item{shape}{'SpatialPolygonsDataFrame' object with the attribute to be checked.}

\item{few.pts}{Logical. If TRUE (default), removes species with few records.}

\item{value}{Minimum number of records to keep the species.}

\item{plot}{Logical. If TRUE (default), the points will be plotted on the provided vector object}

\item{file}{Output filename}
}
\value{
a data frame
}
\description{
A function to mark occurrences based on user supplied vector file.
}
\details{
empty yet
}
\examples{

eug <- filt.generic(Eugenia_aurata, shape = world[world$LONG_NAME == 'Brazil',])
eug

}
\author{
Diogo S. B. Rocha
}
