% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtro_AM.R
\name{filt.generic}
\alias{filt.generic}
\title{Marks occurrences based on user supplied vector file}
\usage{
filt.generic(pts, shape = NULL, few.pts = T, value = 10,
  plot = TRUE, file = NULL)
}
\arguments{
\item{pts}{data.frame. Table with points of occurrence.}

\item{shape}{logical. If TRUE (default), it will check if longitude and latitude are changed. For now this option may be slow when there are many records of occurrence.}

\item{few.pts}{.}

\item{file}{}
}
\value{
a data frame
}
\description{
A function to mark occurrences based on user supplied vector file.
}
\details{

}
\examples{

filt.generic(Eugenia_aurata, shape = br_mun)

}
\author{
Diogo S. B. Rocha
}
