% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtro_AM.R
\name{filt.generic}
\alias{filt.generic}
\title{Marks occurrences based on user supplied vector file}
\usage{
filt.generic(pts, shape = NULL, few.pts = F, value = 10,
  plot = TRUE, file = NULL)
}
\arguments{
\item{pts}{table with points of occurrence.}

\item{shape}{vector object with the attribute to be checked.}

\item{few.pts}{logical. If TRUE (default), removes species with few records.}

\item{value}{minimum number of records to keep the species.}

\item{plot}{logical. If TRUE (default), the points will be plotted on the provided vector object}

\item{file}{Output filename}
}
\value{
a data frame
}
\description{
A function to mark occurrences based on user supplied vector file.
}
\details{
empty yet
}
\examples{

filt.generic(Eugenia_aurata, shape = world)

}
\author{
Diogo S. B. Rocha
}
