% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{filt}
\alias{filt}
\title{Mark occurrences with the name informed different from the coordinate in a user supplied vector file}
\usage{
filt(pts, shape = NULL, field.pts = "municipality",
  field.shape = "NOMEMUNICP")
}
\arguments{
\item{pts}{Data.frame. Table with points of occurrence, including the municipalities informed on the label.}

\item{shape}{It can be a shape of municipalities of Brazil in format "SpatialPolygonsDataFrame".}

\item{field.pts}{String. Name of pts column containing the standard to be searched for.}

\item{field.shape}{String.Name of shape column containing the standard to be searched for.}
}
\value{
a data frame
}
\description{
A function to check whether the coordinates are consistent with the name of the polygon.
}
\details{
The initial idea of this function is to verify if the coordinates informed in the label of the registries are located within the limits of the municipality informed by the label. Returns a table equal to the one entered by adding a column (status) indicating whether the coordinates are consistent.
}
\examples{

eug <- filt(Eugenia_aurata, shape = world, field.pts = "country", field.shape = "LONG_NAME")
eug

}
\author{
Diogo S. B. Rocha
}
