% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fora_muni.R
\name{filt}
\alias{filt}
\title{Mark occurrences with the municipality informed different from the coordinate}
\usage{
filt(pts, shape.municipios = NULL)
}
\arguments{
\item{pts}{data.frame. Table with points of occurrence, including the municipalities informed on the label.}

\item{shape.municipios}{It can be a shape of municipalities of Brazil in format "SpatialPolygonsDataFrame". If it is NULL, the Brazilian shape will be used available on the IBGE website.}
}
\value{

}
\description{
A function to mark occurrences with the municipality informed different from the coordinate.
}
\details{

}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )

}
\seealso{
\code{\link[dismo]{bioclim}}
}
\author{
Diogo S. B. Rocha
}
