% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fora_muni.R
\name{filt}
\alias{filt}
\title{Mark occurrences with the municipality informed different from the coordinate}
\usage{
filt(pts, shape.municipios = NULL)
}
\arguments{
\item{pts}{}

\item{shape.municipios}{number of partitions. The default is 3.}
}
\value{

}
\description{
A function to mark occurrences with the municipality informed different from the coordinate.
}
\details{

}
\examples{
fnames <- list.files(path=paste(system.file(package="dismo"), '/ex', sep=''), pattern='grd', full.names=TRUE )

}
\seealso{
\code{\link[dismo]{bioclim}}
}
\author{
Diogo S. B. Rocha
}
