% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{filt}
\alias{filt}
\title{Mark occurrences with the municipality informed different from the coordinate}
\usage{
filt(pts, shape.municipios = NULL, field.pts = "municipality",
  field.shape = "NOMEMUNICP")
}
\arguments{
\item{pts}{data.frame. Table with points of occurrence, including the municipalities informed on the label. the data frame must contain the following columns in this order: "species","lon","lat", "municipality", "adm1"}

\item{shape.municipios}{It can be a shape of municipalities of Brazil in format "SpatialPolygonsDataFrame". If it is NULL, the Brazilian shape will be used available on the IBGE website.}

\item{field.pts}{string.}

\item{field.shape}{string.}
}
\value{
a data frame
}
\description{
A function to mark occurrences with the municipality informed different from the coordinate.
}
\details{
empty yet
}
\examples{

filt(Eugenia_aurata)

}
\author{
Diogo S. B. Rocha
}
