% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.pts.R
\name{remove.pts}
\alias{remove.pts}
\title{Removing species with few records.}
\usage{
remove.pts(pts, especies, value = 10)
}
\arguments{
\item{pts}{data.frame. Table with points of occurrence.}

\item{especies}{vector.}

\item{value}{Minimum number of records per species.}
}
\value{
a data frame
}
\description{
A function to remove species with few records.
}
\details{
empty yet
}
\examples{

remove.pts(euterpe, unique(euterpe$sp))

}
\author{
Diogo S. B. Rocha
}
